chore: move conversation selector to separate file
