+++ src/reducer/conversationSlice.js
@@ -1,9 +1,9 @@
-import { createSlice, createEntityAdapter, createDraftSafeSelector } from '@reduxjs/toolkit';
+import { createSlice, createEntityAdapter } from '@reduxjs/toolkit';
 const lodashFilter = require('lodash.filter');
 import actions from './conversationSlice.action';
-import { CONVERSATION_PRIORITY_ORDER, MESSAGE_TYPES } from 'constants';
+import { MESSAGE_TYPES } from 'constants';
 import { isEmptyObject } from 'helpers';
-import { applyFilters, findPendingMessageIndex } from '../helpers/conversationHelpers';
+import { findPendingMessageIndex } from '../helpers/conversationHelpers';
 export const conversationAdapter = createEntityAdapter({
   selectId: conversation => conversation.id,
 });
@@ -245,60 +245,7 @@ export const selectConversationToggleStatus = state =>
   state.conversations.isChangingConversationStatus;
 export const selectConversationAssigneeStatus = state =>
   state.conversations.isChangingConversationAssignee;
-export const selectors = {
-  getFilteredConversations: createDraftSafeSelector(
-    [conversationSelector.selectAll, (_, filters) => filters],
-    (conversations, filters) => {
-      const { assigneeType, userId, sortBy } = filters;
-      const comparator = {
-        latest: (a, b) => b.last_activity_at - a.last_activity_at,
-        sort_on_created_at: (a, b) => a.created_at - b.created_at,
-        sort_on_priority: (a, b) => {
-          return CONVERSATION_PRIORITY_ORDER[a.priority] - CONVERSATION_PRIORITY_ORDER[b.priority];
-        },
-      };
-      const sortedConversations = conversations.sort(comparator[sortBy]);
 
-      if (assigneeType === 'mine') {
-        return sortedConversations.filter(conversation => {
-          const { assignee } = conversation.meta;
-          const shouldFilter = applyFilters(conversation, filters);
-          const isAssignedToMe = assignee && assignee.id === userId;
-          const isChatMine = isAssignedToMe && shouldFilter;
-          return isChatMine;
-        });
-      }
-      if (assigneeType === 'unassigned') {
-        return sortedConversations.filter(conversation => {
-          const isUnAssigned = !conversation.meta.assignee;
-          const shouldFilter = applyFilters(conversation, filters);
-          return isUnAssigned && shouldFilter;
-        });
-      }
-
-      return sortedConversations.filter(conversation => {
-        const shouldFilter = applyFilters(conversation, filters);
-        return shouldFilter;
-      });
-    },
-  ),
-  getMessagesByConversationId: createDraftSafeSelector(
-    [conversationSelector.selectEntities, (_, conversationId) => conversationId],
-    (conversations, conversationId) => {
-      const conversation = conversations[conversationId];
-      if (!conversation) {
-        return [];
-      }
-      return conversation.messages;
-    },
-  ),
-  getConversationById: createDraftSafeSelector(
-    [conversationSelector.selectEntities, (_, conversationId) => conversationId],
-    (conversations, conversationId) => {
-      return conversations[conversationId];
-    },
-  ),
-};
 export const {
   clearAllConversations,
   clearConversation,
